requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
rsa==4.9
setuptools==58.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
pytest-asyncio==0.25.2

python-jose==3.3.0
gspread
asyncpg==0.29.0
psycopg2==2.9.9
pytz
pytest
httpx
pytest
pyjwt
python-dotenv
pydantic[email]

